Rename project and reorganize
- Closes [Breaking] Rename Project and Reorganize Commands #16
- Updates `project_patcher` to `prjman`
- Reformats commands under `prjman patch`
    - Will update command names at a later point
