+++ src/prjman/struct/codec.py
