+++ src/prjman/struct/registry.py
